refactor: split DAM direct-access consent and MCP modules
